Convert Hastable to String for Tag Value
In Azure if you export Tag value into a CSV file, it will not be in a readable format. So what you have to do is to convert hastable into string format for all the tag value and then to export it.
The script will help you to do the same.